fix(bbox-adjust): fix bbox measurement
